Clean codes
- Separate Optimizer module.
- Refactor SumMag functions(remove "optimizer" argument)
